Create Python Code for AWS Lambda
The purpose of this code is to trigger a Lambda function once an NBA csv file is uploaded. When the file is uploaded into an S3 bucket, the code will execute and perform parsing of the data, followed by aggregation of players' salary. Once this aggregation is made, it will be displayed in the CloudWatch logs, along with the creation of a file in the same S3 bucket that sends message of success.